refactor(plugin): decompose FlipFinderPanel; lift smart-sell pricing out of UI
Part of #732 (epic #728).

- recommend/SmartSellPricer: smart-sell pricing logic lifted out of the
  Swing layer (calculateSmartSellPrice, loss-minimizing switch, min-
  profitable price, thresholds) as pure static methods.
- ui/panel/PanelFormat: pure presentation helpers (GP/risk/liquidity/
  margin formatting, chart/block icon drawing, colors).
- ui/panel/CardWidgets: shared low-level card builders.
- ui/panel/LoginPanel: the login/auth UI subsystem (login/signup/discord
  device-auth + polling, credential persistence, status) now owns its own
  timers; FlipFinderPanel delegates and reads auth state from the api
  client (removed the duplicated local auth flag).

FlipFinderPanel 4345 -> 3331 lines. Behavior preserved; no runtime
verification possible (Swing) - needs in-game QA. Card-rendering factory
extraction (recommendation/active/completed cards) deferred as a follow-up
since it is woven into focus/blocklist/async view state best split with
in-game iteration.
